+++ injected/src/features/duckplayer/video-overlay.js
@@ -25,19 +25,21 @@
  *   - if the user previously clicked 'watch here + remember', just add the small dax
  *   - otherwise, stop the video playing + append our overlay
  */
-import { SideEffects, VideoParams } from './util.js';
+import { SideEffects, VideoParams, appendImageAsBackground } from './util.js';
 import { DDGVideoOverlay } from './components/ddg-video-overlay.js';
 import { OpenInDuckPlayerMsg, Pixel } from './overlay-messages.js';
 import { IconOverlay } from './icon-overlay.js';
 import { mobileStrings } from './text.js';
 import { DDGVideoOverlayMobile } from './components/ddg-video-overlay-mobile.js';
+import { DDGVideoOverlayMobileAlt } from './components/ddg-video-overlay-mobile-alt.js';
+import { DDGVideoDrawerMobile } from './components/ddg-video-drawer-mobile.js';
 
 /**
  * Handle the switch between small & large overlays
  * + conduct any communications
  */
 export class VideoOverlay {
-    sideEffects = new SideEffects();
+    sideEffects = new SideEffects({ debug: true }); /* TODO: REMOVE debug */
 
     /** @type {string | null} */
     lastVideoId = null;
@@ -178,7 +180,8 @@ export class VideoOverlay {
              */
             const videoElement = document.querySelector(this.settings.selectors.videoElement);
             const playerContainer = document.querySelector(this.settings.selectors.videoElementContainer);
-            if (!videoElement || !playerContainer) {
+            const overlayContainer = document.body; // TODO: Move to RC
+            if (!videoElement || !playerContainer || !overlayContainer) {
                 return null;
             }
 
@@ -217,7 +220,7 @@ export class VideoOverlay {
 
                 // if we get here, we're trying to prevent the video playing
                 this.stopVideoFromPlaying();
-                this.appendOverlayToPage(playerContainer, params);
+                this.appendOverlayToPage(playerContainer, overlayContainer, params);
             }
         }
     }
@@ -226,13 +229,49 @@ export class VideoOverlay {
      * @param {Element} targetElement
      * @param {import("./util").VideoParams} params
      */
-    appendOverlayToPage(targetElement, params) {
+    appendOverlayToPage(targetElement, overlayElement, params) {
         this.sideEffects.add(`appending ${DDGVideoOverlay.CUSTOM_TAG_NAME} or ${DDGVideoOverlayMobile.CUSTOM_TAG_NAME} to the page`, () => {
             this.messages.sendPixel(new Pixel({ name: 'overlay' }));
             const controller = new AbortController();
             const { environment } = this;
 
-            if (this.environment.layout === 'mobile') {
+            /* TODO: remote config */
+
+            /** @type {'a'|'b'} */
+            const mobileLayout = 'b';
+
+            if (this.environment.layout === 'mobile' && mobileLayout === 'a') {
+                const elem = /** @type {DDGVideoOverlayMobileAlt} */ (document.createElement(DDGVideoOverlayMobile.CUSTOM_TAG_NAME));
+                elem.testMode = this.environment.isTestMode();
+                elem.text = mobileStrings(this.environment.strings);
+                elem.addEventListener(DDGVideoOverlayMobileAlt.OPEN_INFO, () => this.messages.openInfo());
+                elem.addEventListener(DDGVideoOverlayMobileAlt.OPT_OUT, (/** @type {CustomEvent<{remember: boolean}>} */ e) => {
+                    return this.mobileOptOut(e.detail.remember).catch(console.error);
+                });
+                elem.addEventListener(DDGVideoOverlayMobileAlt.OPT_IN, (/** @type {CustomEvent<{remember: boolean}>} */ e) => {
+                    return this.mobileOptIn(e.detail.remember, params).catch(console.error);
+                });
+                targetElement.appendChild(elem);
+
+                const drawer = /** @type {DDGVideoDrawerMobile} */ (document.createElement(DDGVideoDrawerMobile.CUSTOM_TAG_NAME));
+                drawer.testMode = this.environment.isTestMode();
+                drawer.text = mobileStrings(this.environment.strings);
+                drawer.addEventListener(DDGVideoDrawerMobile.OPEN_INFO, () => this.messages.openInfo());
+                drawer.addEventListener(DDGVideoDrawerMobile.OPT_OUT, (/** @type {CustomEvent<{remember: boolean}>} */ e) => {
+                    return this.mobileOptOut(e.detail.remember).catch(console.error);
+                });
+                drawer.addEventListener(DDGVideoDrawerMobile.DISMISS, (/** @type {CustomEvent<{remember: boolean}>} */ e) => {
+                    return this.mobileOptOut(false).catch(console.error); // Dismissal should not persist user's choice. Ignore remember-me value.
+                });
+                drawer.addEventListener(DDGVideoDrawerMobile.OPT_IN, (/** @type {CustomEvent<{remember: boolean}>} */ e) => {
+                    return this.mobileOptIn(e.detail.remember, params).catch(console.error);
+                });
+                overlayElement.appendChild(drawer);
+
+                if (elem.container) {
+                    this.appendThumbnail(elem.container);
+                }
+            } else if (this.environment.layout === 'mobile') {
                 const elem = /** @type {DDGVideoOverlayMobile} */ (document.createElement(DDGVideoOverlayMobile.CUSTOM_TAG_NAME));
                 elem.testMode = this.environment.isTestMode();
                 elem.text = mobileStrings(this.environment.strings);
@@ -296,8 +335,15 @@ export class VideoOverlay {
         });
     }
 
+    /**
+     * @param {HTMLElement} overlayElement
+     */
     appendThumbnail(overlayElement) {
-        console.log('TODO', overlayElement);
+        const params = VideoParams.forWatchPage(this.environment.getPlayerPageHref());
+        const videoId = params?.id;
+
+        const imageUrl = this.environment.getLargeThumbnailSrc(videoId);
+        appendImageAsBackground(overlayElement, '.ddg-vpo-bg', imageUrl);
     }
 
     /**
